refactor: colocate NCEP microwave decoding
