+++ earth2studio/data/nnja.py
@@ -36,12 +36,12 @@ import pandas as pd
 import s3fs  # type: ignore[import-untyped]
 from loguru import logger
 
-from earth2studio.data._ncep_microwave import (
+from earth2studio.data.ncep_obs import (
     _NCEP_MICROWAVE_PUBLIC_SCHEMA,
     _NCEP_MICROWAVE_SATELLITES,
     _NCEPMicrowaveAdapter,
+    _NCEPObsSourceBase,
 )
-from earth2studio.data.ncep_obs import _NCEPObsSourceBase
 from earth2studio.data.utils import (
     async_retry,
     datasource_cache_root,
